\name{noise-package}
\alias{noise-package}
\alias{noise}
\docType{package}
\title{
\packageTitle{noise}
}
\description{
\packageDescription{noise}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{noise}
\packageIndices{noise}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{noise}

Maintainer: \packageMaintainer{noise}
}
\references{
Fu, A. Q. and Pachter, L. (2016). Estimating intrinsic and extrinsic noise from single-cell gene expression measurements.  arXiv:1601.03334.
}
\keyword{ package }
\seealso{
\code{\link{elowitz_data}}
\code{\link{yang_nl10}}
\code{\link{computeIntrinsicNoise}}
\code{\link{computeExtrinsicNoise}}
\code{\link{simulateSC}}
}
